Isolating the logic to get the mainnet addresses.
As a safety precaution, we're isolating this logic to environments where the connected wallet is using a testnet.
